requirements.txt
# Core Backend
fastapi
uvicorn[standard]
pydantic
python-multipart

# Communication
paho-mqtt
websockets

# AI & Cognitive (PRISM BRAIN)
numpy
scipy
sounddevice
pyaudio
chromadb
mlx
# openai-whisper

# Utilities
python-dotenv
requests
loguru
pyyaml
